change layout

+++ Makefile
@@ -1,12 +1,17 @@
 
 .PHONEY: clean
 
-seed.iso: seedconfig seedconfig/meta-data seedconfig/user-data
-	[ -f seed.iso ] && rm -f seed.iso && \
-	hdiutil makehybrid -o seed.iso \
-	  -hfs -joliet -iso \
-	  -default-volume-name cidata \
-	  seedconfig
+seed.iso: build/meta-data build/user-data
+	[ -f seed.iso ] && rm -f seed.iso; \
+	hdiutil makehybrid -o seed.iso -hfs -joliet -iso -default-volume-name cidata build 
+
+build/meta-data: meta-data
+	mkdir -p build; \
+	cp meta-data build/meta-data
+
+build/user-data: user-data
+	mkdir -p build; \
+	cp user-data build/user-data
 
 clean:
-	rm -f seed.iso
+	rm -rf build seed.iso
